Feature: Add a space black list
* addresses https://github.com/pacphi/cf-butler/issues/358
