+++ src/main/java/io/pivotal/cfapp/domain/HasOrganizationWhiteList.java
@@ -1,0 +1,7 @@
+package io.pivotal.cfapp.domain;
+
+import java.util.Set;
+
+public interface HasOrganizationWhiteList {
+    Set<String> getOrganizationWhiteList();
+}
