+++ src/main/java/io/pivotal/cfapp/task/StopAppPolicyExecutorTask.java
@@ -1,24 +1,20 @@
 package io.pivotal.cfapp.task;
 
 import java.time.LocalDateTime;
-import java.util.HashSet;
-import java.util.Set;
 
 import org.cloudfoundry.client.v3.applications.StopApplicationRequest;
 import org.cloudfoundry.operations.DefaultCloudFoundryOperations;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
-import org.springframework.util.CollectionUtils;
 
-import io.pivotal.cfapp.config.PasSettings;
 import io.pivotal.cfapp.domain.AppDetail;
 import io.pivotal.cfapp.domain.ApplicationOperation;
-import io.pivotal.cfapp.domain.ApplicationPolicy;
 import io.pivotal.cfapp.domain.HistoricalRecord;
 import io.pivotal.cfapp.service.AppDetailService;
 import io.pivotal.cfapp.service.HistoricalRecordService;
 import io.pivotal.cfapp.service.PoliciesService;
+import io.pivotal.cfapp.util.PolicyFilter;
 import lombok.extern.slf4j.Slf4j;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
@@ -27,7 +23,7 @@ import reactor.core.publisher.Mono;
 @Component
 public class StopAppPolicyExecutorTask implements PolicyExecutorTask {
 
-    private PasSettings settings;
+    private PolicyFilter filter;
     private DefaultCloudFoundryOperations opsClient;
     private AppDetailService appInfoService;
     private PoliciesService policiesService;
@@ -35,13 +31,13 @@ public class StopAppPolicyExecutorTask implements PolicyExecutorTask {
 
     @Autowired
     public StopAppPolicyExecutorTask(
-            PasSettings settings,
+            PolicyFilter filter,
             DefaultCloudFoundryOperations opsClient,
             AppDetailService appInfoService,
             PoliciesService policiesService,
             HistoricalRecordService historicalRecordService
             ) {
-        this.settings = settings;
+        this.filter = filter;
         this.opsClient = opsClient;
         this.appInfoService = appInfoService;
         this.policiesService = policiesService;
@@ -62,20 +58,6 @@ public class StopAppPolicyExecutorTask implements PolicyExecutorTask {
                     log.error("StopAppPolicyExecutorTask terminated with error", error);
                 }
                 );
-    }
-
-    private boolean isBlacklisted(String  organization) {
-        return !settings.getOrganizationBlackList().contains(organization);
-    }
-
-    private boolean isWhitelisted(ApplicationPolicy policy, String organization) {
-        Set<String> prunedSet = new HashSet<>(policy.getOrganizationWhiteList());
-        while (prunedSet.remove(""));
-        Set<String> whitelist =
-                CollectionUtils.isEmpty(prunedSet) ?
-                        prunedSet: policy.getOrganizationWhiteList();
-        return
-                whitelist.isEmpty() ? true: policy.getOrganizationWhiteList().contains(organization);
     }
 
     @Scheduled(cron = "${cron.execution}")
@@ -115,8 +97,8 @@ public class StopAppPolicyExecutorTask implements PolicyExecutorTask {
                 .flatMap(p -> Flux.fromIterable(p.getApplicationPolicies()))
                 .flatMap(ap -> Flux.concat(appInfoService.findByApplicationPolicy(ap, false), appInfoService.findByApplicationPolicy(ap, true)))
                 .distinct()
-                .filter(wl -> isWhitelisted(wl.getT2(), wl.getT1().getOrganization()))
-                .filter(bl -> isBlacklisted(bl.getT1().getOrganization()))
+                .filter(wl -> filter.isWhitelisted(wl.getT2(), wl.getT1().getOrganization()))
+                .filter(bl -> filter.isBlacklisted(bl.getT1().getOrganization(), bl.getT1().getSpace()))
                 .flatMap(ad -> stopApplication(ad.getT1()))
                 .flatMap(historicalRecordService::save);
     }
