+++ src/main/resources/db/mysql/schema.ddl
@@ -11,6 +11,6 @@ CREATE TABLE IF NOT EXISTS legacy_policy ( pk BIGINT NOT NULL AUTO_INCREMENT, id
 CREATE TABLE IF NOT EXISTS application_relationship ( pk BIGINT NOT NULL AUTO_INCREMENT, organization VARCHAR(100), space VARCHAR(100), app_id VARCHAR(50), app_name VARCHAR(100), service_instance_id VARCHAR(50), service_name VARCHAR(100),  service_offering VARCHAR(100), service_plan VARCHAR(50), service_type VARCHAR(30), PRIMARY KEY(pk) );
 CREATE TABLE IF NOT EXISTS historical_record ( pk BIGINT NOT NULL AUTO_INCREMENT, transaction_date_time DATETIME, action_taken VARCHAR(20), organization VARCHAR(100), space VARCHAR(100), app_id VARCHAR(50), service_instance_id VARCHAR(50), type VARCHAR(20), name VARCHAR(300), PRIMARY KEY(pk) );
 CREATE TABLE IF NOT EXISTS space_users ( pk BIGINT NOT NULL AUTO_INCREMENT, organization varchar(100), space varchar(100), auditors VARCHAR(2000), managers VARCHAR(2000), developers VARCHAR(2000), PRIMARY KEY(pk) );
-CREATE TABLE IF NOT EXISTS time_keeper ( collection_time DATETIME );
+CREATE TABLE IF NOT EXISTS time_keeper ( collection_time DATETIME, PRIMARY KEY(collection_time) );
 CREATE OR REPLACE VIEW service_bindings AS select ad.pk, ad.organization, ad.space, ad.app_id, ar.service_instance_id, ad.app_name, ad.buildpack, ad.buildpack_version, ad.image, ad.stack, ad.running_instances, ad.total_instances, ad.memory_used, ad.disk_used, ad.urls, ad.last_pushed, ad.last_event, ad.last_event_actor, ad.last_event_time, ad.requested_state from application_detail ad left join application_relationship ar on ad.app_id = ar.app_id;
 CREATE TABLE IF NOT EXISTS resource_notification_policy ( pk BIGINT NOT NULL AUTO_INCREMENT, id VARCHAR(50), resource_email_template VARCHAR(2000), resource_email_metadata VARCHAR(2000), resource_whitelist VARCHAR(2000), resource_blacklist VARCHAR(2000) , PRIMARY KEY(pk) );
