+++ src/main/resources/db/postgresql/schema.ddl
@@ -11,6 +11,6 @@ CREATE TABLE IF NOT EXISTS legacy_policy ( pk BIGSERIAL PRIMARY KEY, id VARCHAR(
 CREATE TABLE IF NOT EXISTS application_relationship ( pk BIGSERIAL PRIMARY KEY, organization VARCHAR(100), space VARCHAR(100), app_id VARCHAR(50), app_name VARCHAR(100), service_instance_id VARCHAR(50), service_name VARCHAR(100), service_offering VARCHAR(100), service_plan VARCHAR(50), service_type VARCHAR(30) );
 CREATE TABLE IF NOT EXISTS historical_record ( pk BIGSERIAL PRIMARY KEY, transaction_date_time TIMESTAMP, action_taken VARCHAR(20), organization VARCHAR(100), space VARCHAR(100), app_id VARCHAR(50), service_instance_id VARCHAR(50), type VARCHAR(20), name VARCHAR(300) );
 CREATE TABLE IF NOT EXISTS space_users ( pk BIGSERIAL PRIMARY KEY, organization varchar(100), space varchar(100), auditors VARCHAR(512000), managers VARCHAR(512000), developers VARCHAR(512000) );
-CREATE TABLE IF NOT EXISTS time_keeper ( collection_time TIMESTAMP );
+CREATE TABLE IF NOT EXISTS time_keeper ( collection_time TIMESTAMP PRIMARY KEY );
 CREATE OR REPLACE VIEW service_bindings AS select ad.pk, ad.organization, ad.space, ad.app_id, ar.service_instance_id, ad.app_name, ad.buildpack, ad.buildpack_version, ad.image, ad.stack, ad.running_instances, ad.total_instances, ad.memory_used, ad.disk_used, ad.urls, ad.last_pushed, ad.last_event, ad.last_event_actor, ad.last_event_time, ad.requested_state from application_detail ad left join application_relationship ar on ad.app_id = ar.app_id;
 CREATE TABLE IF NOT EXISTS resource_notification_policy ( pk BIGSERIAL PRIMARY KEY, id VARCHAR(50), resource_email_template VARCHAR(512000), resource_email_metadata VARCHAR(512000), resource_whitelist VARCHAR(512000), resource_blacklist VARCHAR(512000) );
